# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Schedule",
        "description": "Documentation de l'API Schedule",
        "contact": {
            "email": "wwwmbassiloic@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API "
        }
    ],
    "paths": {
...
        "/permissions": {
            "get": {
                "tags": [
                    "Permission"
                ],
                "summary": "Get list of permissions",
                "description": "Returns list of permissions",
                "operationId": "getpermissionsList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Permission"
                ],
                "summary": "Store new permission",
                "description": "Returns permission data",
                "operationId": "storepermission",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/permission/{id}": {
            "get": {
                "tags": [
                    "Permission"
                ],
                "summary": "Get permission information",
                "description": "Returns permission data",
                "operationId": "4cb355e0271c8c684ca6f5f9a6a973c6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "permission id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/permission/delete/{id}": {
            "delete": {
                "tags": [
                    "Permission"
                ],
                "summary": "Delete existing permission",
                "description": "Deletes a record and returns no content",
                "operationId": "78787452a6e9560304171ed04aa0f43e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "permission id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/permission/update/{id}": {
            "put": {
                "tags": [
                    "Permission"
                ],
                "summary": "Update existing permission",
                "description": "Returns updated permission data",
                "operationId": "079e22a25ff2fc073436af1c6cea814d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "permission id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "202": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
...
}